requirements.txt
Flask==2.3.2
cryptography==41.0.3
PyJWT==2.8.0
pytest==7.4.0
pytest-cov==4.1.0
requests==2.31.0
flake8==6.1.0
